separate skiplist to discrete files
