+++ src/Makefile
@@ -559,7 +559,8 @@ ENGINE_SERVER_OBJ = \
     vset.o \
     ziplist.o \
     zipmap.o \
-    zmalloc.o
+    zmalloc.o \
+    zskiplist.o
 ENGINE_SERVER_OBJ+=$(ENGINE_TRACE_OBJ)
 ENGINE_CLI_NAME=$(ENGINE_NAME)-cli$(PROG_SUFFIX)
 ENGINE_CLI_OBJ = \
